+++ src/api/middleware/auth.ts
@@ -1,8 +1,8 @@
 import { Request, Response, NextFunction } from "express";
-import { BadRequestError, ForbiddenError, UnauthorizedError } from "../../utils/errors";
-import { isUuidV4 } from "../../utils";
+import { BadRequestError, ForbiddenError, UnauthorizedError } from "../utils/errors";
+import { isUuidV4 } from "../utils";
 
-import type { UserId } from "../../profile/profile.type";
+import type { UserId } from "../validators/profile.type";
 
 const authValidator = async (userId: UserId) => {
   // harcoded for now and returns true if userId is UUID
